+++ frontmark/__about__.py
@@ -1,2 +1,2 @@
-__version__ = '1.0.2.dev'
+__version__ = '1.1.0'
 __description__ = 'CommonMark/Frontmatter reader for Pelican'